Refactor to smaller files
