+++ packages/components/src/Toast/index.ts
@@ -1,1 +1,1 @@
-export { showToast } from "./Toast";
+export { showToast } from "./showToast";
